+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/PasswordInput/PasswordInputConfigurationView.swift
@@ -243,7 +243,7 @@ struct PasswordInputConfigurationView: View {
                                    chips: CircularProgressIndicatorConfigurationModel.Variant.chips)
 
                     if configurationModel.progressVariant == .determinate {
-                        ProgressControl(progress: $configurationModel.progressValue)
+                        DesignToolboxProgressControl(progress: $configurationModel.progressValue)
                     }
                 }